requirements.txt
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-experimental==0.0.53
langchain-text-splitters==0.0.1
openai==0.27.8
SQLAlchemy==2.0.22
pydantic>=1.8.0
pydantic_core==2.10.1
pandas==2.1.3
openpyxl==3.1.2
data
matplotlib
datasets==2.15.0
streamlit==1.32.0
mysql-connector-python==8.3.0
pymysql==1.1.0
tabulate==0.9.0
seaborn==0.13.2
h2o
langchain-openai